+++ packages/auto_mappr_annotation/CHANGELOG.md
@@ -1,8 +1,9 @@
-## 1.0.1
+## 1.0.2
+- Remove `auto_mappr` from annotation example.
 
+## 1.0.1
 - Fix pub.dev issues.
 
 ## 1.0.0
-
 - Initial version.
 - Add `@AutoMappr` annotation.